#2343: added confirmation for shutdown for independent setup

+++ zebrunner.sh
@@ -99,6 +99,16 @@
       exit 0 #no need to proceed as nothing was configured
     fi
 
+    if [[ -z ${SHUTDOWN_CONFIRMED} ]] || [[ ${SHUTDOWN_CONFIRMED} -ne 1 ]]; then
+      # ask about confirmation if it is not confirmed in scope of CE
+      echo_warning "Shutdown will erase all settings and data for \"${BASEDIR}\"!"
+      confirm "" "      Do you want to continue?" "n"
+      if [[ $? -eq 0 ]]; then
+        exit
+      fi
+    fi
+
+
     docker-compose --env-file .env -f docker-compose.yml down -v
 
     rm -f .env
